gemini-streamlit-app/requirements.txt
streamlit>=1.33.0
google-generativeai>=0.7.2
Pillow>=10.2.0
